+++ config/deploy.php
@@ -108,7 +108,7 @@ return [
             'public/images/intro',
             'public/images/kolekcie',
             'public/images/uploaded',
-            'public/images/panorama',
+            'public/panorama',
             'public/sitemaps',
             'storage',
             'resources/fonts',