+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file[^1].
 ### Added
 - Laravel Mix config for compiling assets
 
+### Changed
+- collection published by date, updated filter and ordering
+
+## [2.4.1] - 2020-04-09
 ### Fixed
 - add panorama image dir to deployer config